Add Queryable and Relations traits to Model for enhanced querying and relationship management
